combined all the account info into config.json and implemented receive message parsing and send message for dingtalk
