+++ app.py
@@ -11,17 +11,10 @@ from DDLService import DDLService
 from multi_func_reply import Search
 from database import DataBase
 
-Receive = None
-def arg_parser():
-    parser = argparse.ArgumentParser(description="bot_parser")
-    parser.add_argument("--app", type=str, default='dingtalk', help="app you want to deploy your bot on")
-    parser.add_argument('--debug', action='store_true', help='Add this flag if in debug mode,'
-                                                             'the program will use testing accounts')
-    args = parser.parse_args()
-    global Receive
-    Receive = __import__(f'{args.app}_receive')
+device = None
 
-class app:
+
+class App:
     @staticmethod
     def check_scheduled_task():
         """
@@ -34,8 +27,12 @@ class app:
             # 8:00 AM, show ddl today
             if curr_time.hour == 8 and curr_time.minute == 0:
                 ddl_service = DDLService()
-                Receive.send_msg({'msg_type': 'group', 'number': '705716007', 'msg':
-                    f'大家早上好呀, 又是新的一天，来看看今天还有哪些ddl呢>_<\n{ddl_service.process_query("ddl today".split(" "), "0")}'})
+                if device == 'qq':
+                    Receive.send_msg({'msg_type': 'group_notice', 'number': Receive.GROUP_ACCOUNT,
+                                      'msg': Leetcode.display_questions(Leetcode.question_list)})
+                else:
+                    # TODO: add send message in dingtalk format
+                    pass
                 ddl_service.remove_expired_ddl()  # remove expired ddl timely
                 time.sleep(60)
 
@@ -54,16 +51,30 @@ class app:
             # check if it is midnight
             if curr_time.hour == 0 and curr_time.minute == 0:
                 Leetcode.question_list = DataBase.get_question_on_date(curr_time.strftime('%Y-%m-%d'))
-                Receive.send_msg({'msg_type': 'group_notice', 'number': Receive.GROUP_ACCOUNT,
-                                  'msg': Leetcode.display_questions(Leetcode.question_list)})
+                if device == 'qq':
+                    Receive.send_msg({'msg_type': 'group_notice', 'number': Receive.GROUP_ACCOUNT,
+                                      'msg': Leetcode.display_questions(Leetcode.question_list)})
+                else:
+                    # TODO: add send message in dingtalk format
+                    pass
                 time.sleep(60 * 60 * 23 + 60 * 30)  # sleep 23h 30min
             time.sleep(20)  # allow some buffer time.
 
+    @staticmethod
+    def get_data(text):
+        # info to request api
+        # TODO: specify this in config file
+        data = {
+            "appid": "a612dbe7965b53eeb5eaf26edccc8c94",
+            "userid": "sKJAeMs3",
+            "spoken": text,
+        }
+        return data
 
     @staticmethod
     def get_answer(text):
         # get reply for bot api
-        data = Receive.get_data(text)
+        data = App.get_data(text)
         url = 'https://api.ownthink.com/bot'  # API接口
         response = requests.post(url=url, data=data)
         response.encoding = 'utf-8'
@@ -73,116 +84,6 @@ class app:
         except TypeError:
             return "抱歉，我没有理解您的意思，请换一个问题试试？"
         return answer
-
-
-    @staticmethod
-    def rev_private_msg(rev):
-        if rev['raw_message'] == '在吗':
-            qq = rev['sender']['user_id']
-            Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': Receive.reply_msg[random.randint(0, len(Receive.reply_msg) - 1)]})
-        elif rev['raw_message'] == '你在哪':
-            qq = rev['sender']['user_id']
-            Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': '我无处不在'})
-        # TODO: check if there will be index-out-of-bound issue
-        elif rev['raw_message'].split(' ')[0] == '歌词':
-            qq = rev['sender']['user_id']
-            if len(rev['raw_message'].split(' ')) < 2:
-                Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': '请输入：歌曲<空格><歌曲名>来获得歌曲链接哦'})
-            else:
-                song = rev['raw_message'].replace('歌词', '')
-                d = Search()
-                music_id = d.search_song(song)
-                if music_id is None:
-                    Receive.send_msg(
-                        {'msg_type': 'private', 'number': qq, 'msg': '呜呜呜人家找不到嘛，换首歌试试吧'})
-                else:
-                    text = Search.get_lyrics(music_id)
-                    Receive.send_msg(
-                        {'msg_type': 'private', 'number': qq, 'msg': text})
-        elif rev['raw_message'].split(' ')[0] == '歌曲':
-            qq = rev['sender']['user_id']
-            if len(rev['raw_message'].split(' ')) < 2:
-                Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': '请输入：歌曲<空格><歌曲名>来获得歌曲链接哦'})
-            else:
-                song = rev['raw_message'].replace('歌曲', '')
-                d = Search()
-                music_id = d.search_song(song)
-                if music_id is None:
-                    Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': '呜呜呜人家找不到嘛，换首歌试试吧'})
-                else:
-                    Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': f'[CQ:music,type=163,id={music_id}]'})
-        else:
-            qq = rev['sender']['user_id']
-            content = rev['raw_message']
-            if content == '':
-                answer = '根本搞不懂你在讲咩话，说点别的听听啦'
-            else:
-                answer = app.get_answer(content)
-            Receive.send_msg({'msg_type': 'private', 'number': qq, 'msg': answer})
-
-
-
-    @staticmethod
-    def rev_group_msg(rev):
-        group = rev['group_id']
-        if f'[CQ:at,qq={Receive.BOT_ACCOUNT}]' in rev["raw_message"]:
-            qq = rev['sender']['user_id']
-            message_parts = rev['raw_message'].strip().split(' ')
-            if len(message_parts) < 2:
-                Receive.send_msg({'msg_type': 'group', 'number': group, 'msg': '蛤?'})
-                return
-            if message_parts[1] == '在吗':
-                Receive.send_msg(
-                    {'msg_type': 'group', 'number': group,
-                     'msg': Receive.reply_msg[random.randint(0, len(Receive.reply_msg) - 1)]})
-            elif message_parts[1] in ['wait', 'waitlist', 'Wait']:
-                Receive.send_msg({'msg_type': 'group', 'number': group,
-                                  'msg': f"[CQ:at,qq={qq}]\n" + app.get_waitlist(str(qq))})
-            # leetcode feature
-            elif message_parts[1] == 'leet':
-                Receive.send_msg({'msg_type': 'group', 'number': group,
-                                  'msg': f"[CQ:at,qq={qq}]\n" + Leetcode.process_query(message_parts, qq)})
-            # DDL feature
-            elif message_parts[1] == 'ddl':
-                Receive.send_msg({'msg_type': 'group', 'number': group,
-                                  'msg': f"[CQ:at,qq={qq}]\n" + DDLService.process_query(message_parts[1:], qq)})
-            # TODO: add help reply
-            else:
-                content = ""
-                for i in range(1, len(message_parts)):
-                    content += message_parts[i] + " "
-                Receive.send_msg({'msg_type': 'group', 'number': group,
-                                  'msg': f'[CQ:at,qq={qq}]' + Receive.get_answer(content)})
-
-
-    @staticmethod
-    def message_process_tasks():
-        """
-        All private/group message processing are done here.
-        """
-        while True:
-            received = Receive.rev_msg()
-            try:
-                if received["post_type"] == "message":
-                    # private message
-                    if received["message_type"] == "private":
-                        Receive.rev_private_msg(received)
-                    # group message
-                    elif received["message_type"] == "group":
-                        Receive.rev_group_msg(received)
-            # TODO: consider enlarge type of Error?
-            except TypeError as e:
-                # This error will be reported to developers via qq private message.
-                error_msg = '[Internal Error] TypeError while trying to reply message. ' \
-                            'If the message received is too long, try to' + \
-                            'release the length restriction. (currently 4096)\n' + \
-                            f'[Exception Message] {e}\n ' \
-                            f'Message received: {received}'
-                # TODO: specify admin account in config file
-                Receive.send_msg({'msg_type': 'private', 'number': '2220038250', 'msg': error_msg})
-                Receive.send_msg({'msg_type': 'private', 'number': '3429582673', 'msg': error_msg})
-                # also record in log
-                print(f'##### Error\n{error_msg}')
 
     def get_waitlist(qq: str) -> str:
         url = "https://w5.ab.ust.hk/msapi/sis/stdt_class_enrl/%7BstdtID%7D"
@@ -209,3 +110,16 @@ class app:
             info += f'course code: {c["crseCode"]}, waitlist position: {c["waitPosition"]}\n'
 
         return info
+
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser(description="bot_parser")
+    parser.add_argument("--app", type=str, default='dingtalk', help="app you want to deploy your bot on")
+    parser.add_argument('--debug', action='store_true', help='Add this flag if in debug mode,'
+                                                             'the program will use testing accounts')
+    args = parser.parse_args()
+    device = args.app
+    if args.app == 'qq':
+        from qq_receive import Receive
+    else:
+        from dingtalk_receive import Receive
