--- studentAcount.json
@@ -1,4 +1,0 @@
-{
-	"2220038250": "eyJhbGciOiJSUzI1NiIsImtpZCI6ImNhcyJ9.eyJqdGkiOiJTVC0xOTE3Ni1JUnpNWVhvR1lWbG1DNzh4ekRYWkFhT1d4Y2tjYXMxIiwiaXNzIjoiaHR0cHM6Ly9jYXMudXN0LmhrL2Nhcy9vaWRjIiwiYXVkIjoiMjAwMDkiLCJleHAiOjE2NTk5MDA2MjMsImlhdCI6MTY1OTg3MTgyMywibmJmIjoxNjU5ODcxNTIzLCJzdWIiOiJqbGl1ZHEiLCJhbXIiOlsiTGRhcEF1dGhlbnRpY2F0aW9uSGFuZGxlciIsIkhLVVNUIENBUyJdLCJzdGF0ZSI6ImpXc0hQQ2tCdURGWDRVa2NEZml1aVEiLCJub25jZSI6IjcxRHc1c05PWVh3ekFFUEd6bExUM1EiLCJhdF9oYXNoIjoiUHJtUE95cXBZbTZtX1NfNmh6QmxodyIsImRlcGFydG1lbnROdW1iZXIiOiJDU0UiLCJlbWFpbCI6ImpsaXVkcUBjb25uZWN0LnVzdC5oayIsImVtcGxveWVlVHlwZSI6InN0dWRlbnQiLCJuYW1lIjoiTElVLCBKaWFubWVuZyIsIm91IjpbIkFETUlTIFVzZXIiLCJTdHVkZW50IiwiVW5kZXJncmFkdWF0ZSJdLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJqbGl1ZHEifQ.R3AYjLNlrknGVUzcSwLmmciOriI3DZW4NYbSaiMJ-NlXqe7muQmuAyMh-GbeMJVYAtHiDzIIpivHw7SIMYawSRjzONQJZ4f6bEHtvdbA_wu1ZaNFcWJBbaxK9XK47GxOY1-Iuri7-hsPIpmCgUCi2jDip-5SLSFoPhXVhgYq87XoDIi_aLvBzTjquZRiJ48LdTIVtM0DaBsNK7yXd_NDpEo7UyQXQNimygtAJ_Dw3Vyq5PSVJ4B9XetSyFa1n3SrVmviXgVKfO2kdzQDuLbcklhbGXqq9hJxqPBwkX4lWHHChFo6BhNW7JgebZP7xiW5-9RmiHIMYPoJ5aBV0pHk2Q",
-	"3429582673": "eyJhbGciOiJSUzI1NiIsImtpZCI6ImNhcyJ9.eyJqdGkiOiJTVC01MjIxOS10amk4S3F6ME1STUJiQU1pdHhiR2VHMVotUGNjYXMyIiwiaXNzIjoiaHR0cHM6Ly9jYXMudXN0LmhrL2Nhcy9vaWRjIiwiYXVkIjoiMjAwMDkiLCJleHAiOjE2NjE1MjQyMjMsImlhdCI6MTY2MTQ5NTQyMywibmJmIjoxNjYxNDk1MTIzLCJzdWIiOiJ6bWVuZ2FmIiwiYW1yIjpbIkxkYXBBdXRoZW50aWNhdGlvbkhhbmRsZXIiLCJIS1VTVCBDQVMiXSwic3RhdGUiOiJKYlNzdGVOcXlibzh3R25qYVA1dGRBIiwibm9uY2UiOiJ4WktQVzNJZEI3QWpBbTYzVFFqeW93IiwiYXRfaGFzaCI6IkZqX2Z3NzRfWHpmM0xPaXlhSnJRT3ciLCJkZXBhcnRtZW50TnVtYmVyIjoiQ1NFIiwiZW1haWwiOiJ6bWVuZ2FmQGNvbm5lY3QudXN0LmhrIiwiZW1wbG95ZWVUeXBlIjoic3R1ZGVudCIsIm5hbWUiOiJNRU5HLCBaZXl1YW4iLCJvdSI6WyJBRE1JUyBVc2VyIiwiU3R1ZGVudCIsIlVuZGVyZ3JhZHVhdGUiXSwicHJlZmVycmVkX3VzZXJuYW1lIjoiem1lbmdhZiJ9.lH2MVZFmfN-47VQPoTpHB8ae0tSM1rE5aKrD8G_El-LpB_4n960oIxpITa-Ms3uNqlSbHveICwpmoQT0F8gOOm3nxQ5fzBO1I0Xi6XDSsNvnFYXDDNg6pAo5TmNhG6XQO-Vh-5k1NX303mdRO74Qui74uNDdm4DNocNRhTX3IpNmylutZ7VYaGt0S_gYG_G8JYlGdIyJKHnggwMSiUhZ_wUnkquaVT6JjJxSxwzTpdHdfboPaYO74t6MMRwW_5d918hTSOnJkZn51XaTRyn4qQ3OnoJhRn1u8KTdXjaPHdWUI6eKmO_exJ3Sii-SageK0MT3fjZDS0gYBxWuEGtEvw"
-}
